Board

Description

Module for the Board and all its components.

Documentation

data Board

Constructors

Board

Fields

width :: Int

height :: Int

numberOfMines :: Int

fields :: [Field]

gameState :: GameState

data Field

Constructors

Field

Fields

content :: FieldContent

state :: FieldState

data FieldContent

Constructors

Mine

NoMine NeighbourCount

data FieldState

Constructors

Revealed

Hidden Bool

data NeighbourCount

Constructors

Nil

One

Two

Three

Four

Five

Six

Seven

Eight

data GameState

Constructors

GameWon

GameLost

GameNotFinished

board

Arguments

:: Int
the width

-> Int
the height

-> Int
the number of mines

-> Int
the seed

-> Board
the new Board.

Create a new Board.

flagField

Arguments

:: Board
receiver

-> Int
x coordinate

-> Int
y coordinate

-> Board
the new board

Flags the field on the given coordinate.

getFieldContent

Arguments

:: (Int, Int)
the coordinate

-> [(Int, Int)]
the coordinates of the mines

-> Int
the width of the board

-> Int
the height of the board

-> FieldContent

Calculates the number of mines surrounding the field at the given coordinate and determines its content.

isGameLost

Arguments

:: Board
receiver

-> Int
x coordinate

-> Int
y coordinate

-> Bool
is game lost

Tells if a field that is being reveald is a mine and therefore the loosing condtiion is met.

revealBoard

Arguments

:: Board
the receiver

-> Int
the index

-> Board
the changed board

Reveals a field by index like checkedRevealField.

revealField

Arguments

:: Board
receiver

-> Int
x coordinate

-> Int
y coordinate

-> Board
new board

Sets the state of the field on the board with the coordinates of x and y to revealed.

isGameWon

Arguments

:: Board
receiver

-> Bool
is game won

Tells if the board is solved and therefore the winning condition is met.

checkedRevealField

Arguments

:: Board
the receiver

-> Int
x coordinate

-> Int
y coordinate

-> Board
the changed Board

Reveals a field and recursively reveals all nil fields and the first non-nil field except mines. If the selected field is a mine the game state will be changed to lost. If there are no more fields to reveal but mines the game state will be changed to won. Else the game state stays at not finished.

revealRecursive

Arguments

:: Board
receiver

-> [(Int, Int)]
coordinates to reveal

-> Board
the new board

Tries to reveal all given coordinates recursivly. If a field is not a mine and also has the value Nil, then it will also recursively reveal the neighbouring fields that are not diagonally located to this field.

getFieldFromBoard

Arguments

:: Board
receiver

-> Int
x coordinate

-> Int
y coordinate

-> Field
the field

Returns the field on the given coordinate.

getCrossNeighbourCellPositions

Arguments

:: Board
receiver

-> Int
x coordinate

-> Int
y coordinate

-> [(Int, Int)]
coordinates of neighbouring cells that are within the boundary

Returns the coordinates that are above, below, left and right from the given coordinate. Coordinates that are not within the boundary will be filtered out.